не смог настроить конфигурационный файл для работы с аннотацией @Slf4j, для вывода в доп. задании использовал параметр аннотации, который случайно сам обнаружил методом тыка (topic = "TRACE")
