+++ pom.xml
@@ -59,6 +59,12 @@
             <version>3.17.0</version>
         </dependency>
 
+        <dependency>
+            <groupId>org.zalando</groupId>
+            <artifactId>logbook-spring-boot-starter</artifactId>
+            <version>3.7.2</version>
+        </dependency>
+
     </dependencies>
 
     <!-- блок настройки процесса сборки -->
@@ -71,8 +77,8 @@
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>3.11.0</version>
                 <configuration>
-                    <source>22</source>
-                    <target>22</target>
+                    <source>21</source>
+                    <target>21</target>
 
                 </configuration>
             </plugin>
