+++ src/main/resources/application.properties
@@ -1,1 +1,1 @@
-
+logging.level.org.zalando.logbook=TRACE
